remove rubocop offences
