+++ app/controllers/urls_controller.rb
@@ -18,7 +18,7 @@ class UrlsController < ApplicationController
         redirect_to :root
       end
     else
-      @url_in_database = Url.find_by_url_clean(@url.url_clean)
+      @url_in_database = Url.find_by(url_clean: @url.url_clean)
       render js: "document.getElementById('short_url_box').classList
       .remove('d-none');document.getElementById('short_url')
       .innerText='http://#{request.host}/#{@url_in_database.url_short}/';"
@@ -26,7 +26,7 @@ class UrlsController < ApplicationController
   end
 
   def show
-    @url = Url.find_by_url_short(params[:url_short])
+    @url = Url.find_by(url_short: params[:url_short])
     redirect_to "http://#{@url.url_clean}/"
   end
