+++ app/validators/url_validator.rb
@@ -1,10 +1,14 @@
 class UrlValidator < ActiveModel::EachValidator
-    def validate_each(record, attribute, value)
-      record.errors[attribute] << (options[:message] || "must be a valid URL") unless url_valid?(value)    
-    end
-  
-    def url_valid?(url)
-      url = URI.parse(url) rescue false
-      url.kind_of?(URI::HTTP) || url.kind_of?(URI::HTTPS)
-    end 
+  def validate_each(record, attribute, value)
+    record.errors[attribute] << (options[:message] || 'must be a valid URL') unless url_valid?(value)
   end
+
+  def url_valid?(url)
+    begin
+      url = URI.parse(url)
+    rescue URI::InvalidURIError
+      false
+    end
+    url.is_a?(URI::HTTP) || url.is_a?(URI::HTTPS)
+  end
+end
